+++ CHANGELOG.md
@@ -42,6 +42,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   for documentation ([#616](https://github.com/Neoteroi/BlackSheep/issues/616)).
 - Improve the build matrix to build wheels for `arm64` architecture for Linux and
   Windows, and use `cibuildwheel` for Ubuntu and Windows, by @bymoye and @RobertoPrevato.
+- Update type annotations to Python >= 3.10.
+- Fix bug that would prevent union types described using pipes from being properly
+  represented in OpenAPI specification.
+- Add support for alternative programming-style naming for generic types in OpenAPI
+  specification files. When enabled, type names use underscore notation closer to
+  actual type annotations (e.g., `PaginatedSet_Address` instead of
+  `PaginatedSetOfAddress`, `Dict_str_int` instead of `DictOfstrAndint`). This can be
+  controlled via the `programming_names` parameter in `DefaultSerializer` or the
+  `APP_OPENAPI_PROGRAMMING_NAMES` environment variable, setting it to a truthy value
+  ('1' or 'true').
 
 ## [2.4.3] - 2025-10-19 :musical_keyboard: